Updated Queue.h

+++ Queue/Queue.h
@@ -49,7 +49,7 @@ public:
         return elements.empty();
     }
     
-    unsigned long size()
+    size_t size()
     {
         return elements.size();
     }
